refactor: move LoginInput to the components directory
